+++ src/main.rs
@@ -109,9 +109,6 @@ fn linker_main() -> Status {
             (author: "TON Labs")
             (@arg INPUT: +required +takes_value "TVM assembler source file")
             (@arg ABI: -a --("abi-json") +takes_value "Supplies contract abi to calculate correct function ids. If not specified abi can be loaded from file path obtained from <INPUT> path if it exists.")
-            (@arg CTOR_PARAMS: -p --("ctor-params") +takes_value "Supplies arguments for the constructor")
-            (@arg GENKEY: --genkey +takes_value conflicts_with[SETKEY] "Generates new keypair for the contract and saves it to the file")
-            (@arg SETKEY: --setkey +takes_value conflicts_with[GENKEY] "Loads existing keypair from the file")
             (@arg WC: -w +takes_value "Workchain id used to print contract address, -1 by default.")
             (@arg DEBUG: --debug "Prints debug info: xref table and parsed assembler sources")
             (@arg DEBUG_MAP: --("debug-map") +takes_value "Generates debug map file")
@@ -164,13 +161,6 @@ fn linker_main() -> Status {
             (@arg ADDRESS: --addr +takes_value "Optional destination address to support ABI 2.3")
             (@arg INPUT: +required +takes_value "TVM assembler source file or contract name")
         )
-        (@subcommand init =>
-            (about: "initialize smart contract public variables")
-            (version: build_info.as_str())
-            (@arg INPUT: +required +takes_value "Path to compiled smart contract file")
-            (@arg DATA: +required +takes_value "Set of public variables with values in json format")
-            (@arg ABI: +required +takes_value "Path to smart contract ABI file")
-        )
         (@subcommand disasm =>
             (about: "disassemble a tvc or dumps its tree of cells")
             (version: build_info.as_str())
@@ -195,10 +185,6 @@ fn linker_main() -> Status {
         (@setting SubcommandRequired)
     ).get_matches();
 
-    //SUBCOMMAND INIT
-    if let Some(matches) = matches.subcommand_matches("init") {
-        return run_init_subcmd(matches);
-    }
 
     //SUBCOMMAND TEST
     if let Some(test_matches) = matches.subcommand_matches("test") {
@@ -287,20 +273,6 @@ fn linker_main() -> Status {
             ParseEngine::new(sources, abi_json)?
         );
 
-        match compile_matches.value_of("GENKEY") {
-            Some(file) => {
-                let pair = KeypairManager::new();
-                pair.store_public(&(file.to_string() + ".pub"))?;
-                pair.store_secret(file)?;
-                prog.set_keypair(pair.drain());
-            },
-            None => if let Some(file) = compile_matches.value_of("SETKEY") {
-                let pair = KeypairManager::from_secret_file(file)
-                    .ok_or_else(|| format_err!("Failed to read keypair"))?;
-                prog.set_keypair(pair.drain());
-            },
-        };
-
         let debug = compile_matches.is_present("DEBUG");
         prog.set_language(compile_matches.value_of("LANGUAGE"));
 
@@ -312,15 +284,9 @@ fn linker_main() -> Status {
             .map(|wc| wc.parse::<i8>().unwrap_or(-1))
             .unwrap_or(-1);
 
-        let ctor_params = compile_matches.value_of("CTOR_PARAMS");
-        if ctor_params.is_some() && abi_file.is_none() {
-            let msg = "ABI is mandatory when CTOR_PARAMS is specified.";
-            bail!(msg);
-        }
-
         let data_filename = compile_matches.value_of("DATA");
 
-        prog.compile_to_file_ex(wc, abi_file, ctor_params, out_file, debug, data_filename)?;
+        prog.compile_to_file_ex(wc, out_file, data_filename, false)?;
 
         if compile_matches.is_present("DEBUG_MAP") {
             let filename = compile_matches.value_of("DEBUG_MAP").unwrap();
@@ -435,26 +401,6 @@ fn parse_ticktock(ticktock: Option<&str>) -> Result<Option<i8>> {
     } else {
         Ok(None)
     }
-}
-
-fn set_initial_data(tvc: &str, data: &str, abi: &str) -> Status {
-    let abi = load_abi_json_string(abi)?;
-    let mut state_init = load_from_file(tvc)?;
-    let new_data = ton_abi::json_abi::update_contract_data(
-        &abi,
-        data,
-        state_init.data.clone().unwrap_or_default().into(),
-    )?;
-    state_init.set_data(new_data.into_cell());
-    save_to_file(state_init, None, 0)?;
-    Ok(())
-}
-
-fn run_init_subcmd(matches: &ArgMatches) -> Status {
-    let tvc = matches.value_of("INPUT").unwrap();
-    let data = matches.value_of("DATA").unwrap();
-    let abi = matches.value_of("ABI").unwrap();
-    set_initial_data(tvc, data, abi)
 }
 
 fn decode_hex_string(hex_str: String) -> Result<(Vec<u8>, usize)> {
@@ -619,8 +565,7 @@ fn build_body(matches: &ArgMatches, address: Option<String>) -> Result<Option<Sl
     if mask == 0x3 {
         let key_file = match matches.value_of("SIGN") {
             Some(path) => {
-                let pair = KeypairManager::from_secret_file(path)
-                    .ok_or_else(|| format_err!("Failed to read keypair."))?;
+                let pair = KeypairManager::from_file(path)?;
                 Some(pair.drain())
             },
             _ => None